+++ packages/firebase_crashlytics/ios/Classes/FirebaseCrashlyticsPlugin.m
@@ -76,11 +76,16 @@
     for (NSDictionary *errorElement in errorElements) {
       [frames addObject:[self generateFrame:errorElement]];
     }
-    [[Crashlytics sharedInstance]
-        recordCustomExceptionName:call.arguments[@"exception"]
-                           reason:[NSString
-                                      stringWithFormat:@"thrown %s", call.arguments[@"context"]]
-                       frameArray:frames];
+
+    NSString *context = call.arguments[@"context"];
+    NSString *reason;
+    if (context != nil) {
+      reason = [NSString stringWithFormat:@"thrown %@", context];
+    }
+
+    [[Crashlytics sharedInstance] recordCustomExceptionName:call.arguments[@"exception"]
+                                                     reason:reason
+                                                 frameArray:frames];
     result(@"Error reported to Crashlytics.");
   } else if ([@"Crashlytics#isDebuggable" isEqualToString:call.method]) {
     result([NSNumber numberWithBool:[Crashlytics sharedInstance].debugMode]);