+++ packages/firebase_in_app_messaging/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.1+2
+
+* Remove dependency `androidx.annotation:annotation:1.0.0`.
+
 ## 0.0.1+1
 
 * Update google-services Android gradle plugin to 4.3.0 in documentation and examples.